+++ src/index.ts
@@ -1,6 +1,8 @@
 import type { Maidr } from '@type/grammar';
 import { DomEventType } from '@type/event';
+import { MaidrApp } from '@ui/App';
 import { Constant } from '@util/constant';
+import { createRoot } from 'react-dom/client';
 import { Controller } from './controller';
 
 if (document.readyState === 'loading') {
@@ -48,7 +50,6 @@ function main(): void {
 
 function initMaidr(maidr: Maidr, plot: HTMLElement): void {
   let maidrContainer: HTMLElement | null = null;
-  let reactContainer: HTMLElement | null = null;
   let controller: Controller | null = null;
 
   const onFocusOut = (): void => {
@@ -68,16 +69,27 @@ function initMaidr(maidr: Maidr, plot: HTMLElement): void {
   const onFocusIn = (): void => {
     // Allow React to process all the events before focusing in.
     setTimeout(() => {
-      if (!maidrContainer || !reactContainer) {
+      if (!maidrContainer) {
         return;
       }
 
       if (!controller) {
         // Create a deep copy to prevent mutations on the original maidr object.
         const maidrClone = JSON.parse(JSON.stringify(maidr));
-        controller = new Controller(maidrClone, plot, reactContainer);
+        controller = new Controller(maidrClone, plot);
       }
     }, 0);
+  };
+
+  const onVisibilityChange = (): void => {
+    if (document.visibilityState === 'visible') {
+      if (controller) {
+        controller.dispose();
+        controller = null;
+      }
+      const maidrClone = JSON.parse(JSON.stringify(maidr));
+      controller = new Controller(maidrClone, plot);
+    }
   };
 
   const figureElement = document.createElement(Constant.FIGURE);
@@ -90,7 +102,7 @@ function initMaidr(maidr: Maidr, plot: HTMLElement): void {
   figureElement.parentNode!.replaceChild(articleElement, figureElement);
   articleElement.appendChild(figureElement);
 
-  reactContainer = document.createElement(Constant.DIV);
+  const reactContainer = document.createElement(Constant.DIV);
   reactContainer.id = `${Constant.REACT_CONTAINER}-${maidr.id}`;
   figureElement.appendChild(reactContainer);
 
@@ -99,10 +111,15 @@ function initMaidr(maidr: Maidr, plot: HTMLElement): void {
   plot.addEventListener(DomEventType.CLICK, onFocusIn);
   maidrContainer.addEventListener(DomEventType.FOCUS_OUT, onFocusOut);
 
+  document.addEventListener(DomEventType.VISIBILITY_CHANGE, onVisibilityChange);
+
+  const reactRoot = createRoot(reactContainer, { identifierPrefix: maidr.id });
+  reactRoot.render(MaidrApp(plot));
+
   (() => {
     // Create a deep copy to prevent mutations on the original maidr object.
     const maidrClone = JSON.parse(JSON.stringify(maidr));
-    const controller = new Controller(maidrClone, plot, reactContainer);
+    const controller = new Controller(maidrClone, plot);
     controller.dispose();
   })();
 }