+++ src/type/event.ts
@@ -6,7 +6,11 @@ export enum DomEventType {
   DOM_LOADED = 'DOMContentLoaded',
   FOCUS_IN = 'focusin',
   FOCUS_OUT = 'focusout',
+  MOUSE_ENTER = 'mouseenter',
+  MOUSE_LEAVE = 'mouseleave',
+  MOUSE_MOVE = 'mousemove',
   SELECTION_CHANGE = 'selectionchange',
+  VISIBILITY_CHANGE = 'visibilitychange',
 }
 
 export type Status